pythonProject1/service_employee/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
idna==3.7
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
